moved WeaponsCardCustomMag to container folder
